# api.yaml
swagger: "2.0.0"
info:
  title: Academy APP
  description: Api Documentation for Academia Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/auth/loginAdmin:
    post:
      tags:
        - name: Admin Auth
      summary: Login System Admin Test
      description: This api will be for the login of the admin in the test app
      consumes: application/json
      produces: application/json
      parameters:
        - in: body
          name: User
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
...
  /v1/auth/updateNameAdmin/{userID}:
    put:
      tags:
        - name: Admin Auth
      summary: Update Name
      description: Update NAme of the user
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - in: body
          name: data
          schema:
            type: object
            properties:
              name:
                type: string

      responses:
        200:
          description: OK
  /v1/auth/getUsersByFilter/{offset}/{limit}:
    post:
      tags:
        - name: User
      summary: Filters Users
      description: APi will get the filters in the user
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: offset
          in: path
          type: string
          required: true
        - name: limit
          in: path
          type: string
          required: true
        - in: body
          name: data
          schema:
            type: object
            properties:
              school:
                type: string
              status:
                type: boolean
              from:
                type: date
              to:
                type: date

      responses:
        200:
          description: OK
...
  /v1/auth/addProfileAdmin/{userID}:
    put:
      tags:
        - name: Admin Auth
      summary: Update Profile
      description: Updates the profile image
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: imageProfile
          in: formData
          type: file
          required: true
      responses:
        200:
          description: OK
  /v1/auth/updatePasswordAdmin:
    post:
      tags:
        - name: Admin Auth
      summary: Update Password
      description: This api will be used to Update the password in the Database
      consumes: application/json
      produces: application/json
      parameters:
        - in: body
          name: ForgetPassword
          schema:
            type: object
            properties:
              email:
                type: string
              password:
                type: string
              newPassword:
                type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/auth/getUserByID/{userID}:
    get:
      tags:
        - name: User
      summary: Users By ID
      description: GEtting users by there ID
      parameters:
        - name: userID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/auth/getAllUsers:
    get:
      tags:
        - name: User
      summary: Getting all Users
      description: APi will get all the users in the database
      parameters:
        - name: page
          in: query
          type: string
        - name: limit
          in: query
          type: string
        - name: page
          in: query
          type: string
        - name: searchString
          in: query
          type: string
        - name: isActive
          in: query
          type: boolean
        - name: startDate
          in: query
          type: string
        - name: endDate
          in: query
          type: string
        - name: sortByDate
          in: query
          type: string
          enum: [asc, dsc]
      responses:
        200:
          description: OK
  /v1/auth/changeUserStatus/{userID}:
    post:
      tags:
        - name: Admin
      summary: Change user Status
      consumes: multipart/form-data
      produces: application/json
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: isActive
          in: formData
          type: boolean
          required: true
      responses:
        200:
          description: OK
  /v1/auth/deleteUser/{userID}:
    delete:
      tags:
        - name: Admin
      summary: Delete User
      description: Delete user by a particular id
      parameters:
        - name: userID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/board/createBoard:
    post:
      tags:
        - name: Board
      summary: Create Board
      description: Api to Create Board
      consumes: application/json
      produces: application/json
      parameters:
        - name: name
          in: formData
          type: string
          required: true
        - name: status
          in: formData
          type: string
          enum: [active, inactive]
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/Board/editBoard/{boardID}:
    patch:
      tags:
        - name: Board
      summary: Edit the Board
      description: Api that will edit the Board
      consumes: application/json
      produces: application/json
      parameters:
        - name: boardID
          in: path
          type: string
          required: true
        - name: name
          in: formData
          type: string
        - name: status
          in: formData
          type: string
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  /v1/board/deleteBoardById/{boardId}:
    delete:
      tags:
        - name: Board
      summary: Delete Board by its id
      description: APi to delete board by id
      parameters:
        - name: boardId
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
        400:
          description: Invalid Credentials
        404:
          description: Not Found
        401:
          description: Invalid Credentials
        500:
          description: Internal Server Error
  #### ------------- SubBoard Route ------------- ####
...
  /v1/cart/add:
    post:
      tags:
        - Cart
      summary: course add to cart
      parameters:
        - in: body
          name: cart
          schema:
            type: object
            properties:
              userID:
                type: string
              courseID:
                type: string
      responses:
        200:
          description: OK
  /v1/cart/remove:
    patch:
      tags:
        - Cart
      summary: course remove from cart
      parameters:
        - in: body
          name: cart
          schema:
            type: object
            properties:
              userID:
                type: string
              courseID:
                type: string
      responses:
        200:
          description: OK
  /v1/cart/all/{userID}:
    get:
      tags:
        - Cart
      summary: get users cart
      parameters:
        - in: path
          name: userID
          type: string
          required: true
      responses:
        200:
          description: OK
...